+++ tests/suite/dfxvm_init.rs
@@ -2,6 +2,26 @@ use crate::common::{file_contents::manifest_json, ReleaseAsset, ReleaseServer, T
 use assert_cmd::prelude::*;
 use predicates::str::*;
 use std::os::unix::fs::PermissionsExt;
+use std::path::PathBuf;
+
+// This fake rc copied from https://github.com/rust-lang/rustup/blob/master/tests/suite/cli_paths.rs
+// Let's write a fake .rc which looks vaguely like a real script.
+const FAKE_RC: &str = r#"
+# Sources fruity punch.
+. ~/fruit/punch
+
+# Adds apples to PATH.
+export PATH="$HOME/apple/bin"
+"#;
+
+fn posix_source() -> String {
+    #[cfg(target_os = "macos")]
+    let env_path = "$HOME/Library/Application Support/org.dfinity.dfx/env";
+    #[cfg(target_os = "linux")]
+    let env_path = "$HOME/.local/share/dfx/env";
+
+    format!(". \"{env_path}\"\n", env_path = env_path)
+}
 
 #[test]
 fn default_installation() {
@@ -121,7 +141,7 @@ fn specific_dfx_version() {
 #[test]
 fn xdg_data_home_set() {
     let home_dir = TempHomeDir::new();
-    let xdg_data_home = home_dir.path().join(".custom/xdg/data-home");
+    let xdg_data_home = home_dir.join(".custom/xdg/data-home");
     let home_dir = home_dir.with_xdg_data_home(&xdg_data_home);
     let server = ReleaseServer::new(&home_dir);
 
@@ -218,4 +238,496 @@ fn sets_permissions_from(from_mode: u32) {
     // and the installed dfx proxy
     let metadata = std::fs::metadata(home_dir.installed_dfx_proxy_path()).unwrap();
     assert_eq!(metadata.permissions().mode(), MODE_FILE | 0o755);
+}
+
+#[test]
+fn creates_dot_profile() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    home_dir.dfxvm_init().arg("--proceed").assert().success();
+
+    let dot_profile = home_dir.join(".profile");
+    let contents = std::fs::read_to_string(dot_profile).unwrap();
+    assert_eq!(contents, posix_source());
+}
+
+#[test]
+fn updates_dot_profile() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    let dot_profile = home_dir.join(".profile");
+    std::fs::write(&dot_profile, FAKE_RC).unwrap();
+
+    home_dir.dfxvm_init().arg("--proceed").assert().success();
+
+    let expected = FAKE_RC.to_owned() + &posix_source();
+    let new_dot_profile = std::fs::read_to_string(&dot_profile).unwrap();
+    assert_eq!(new_dot_profile, expected);
+}
+
+#[test]
+fn adds_newline_if_existing_file_does_not_end_in_one() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    let dot_profile = home_dir.join(".profile");
+    std::fs::write(&dot_profile, FAKE_RC.trim_end()).unwrap();
+
+    home_dir.dfxvm_init().arg("--proceed").assert().success();
+
+    let expected = FAKE_RC.to_owned() + &posix_source();
+    let new_dot_profile = std::fs::read_to_string(&dot_profile).unwrap();
+    assert_eq!(new_dot_profile, expected);
+}
+
+#[test]
+fn updates_bash_rcs() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    let rcs: Vec<PathBuf> = [".bashrc", ".bash_profile", ".bash_login"]
+        .iter()
+        .map(|rc| home_dir.join(rc))
+        .collect();
+    for rc in &rcs {
+        std::fs::write(rc, FAKE_RC).unwrap();
+    }
+
+    home_dir.dfxvm_init().arg("--proceed").assert().success();
+
+    let expected = FAKE_RC.to_owned() + &posix_source();
+    for rc in &rcs {
+        let new_rc = std::fs::read_to_string(rc).unwrap();
+        assert_eq!(new_rc, expected);
+    }
+}
+
+#[test]
+fn does_not_create_bash_rcs() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+    server.expect_install_latest();
+
+    let rcs: Vec<PathBuf> = [".bashrc", ".bash_profile", ".bash_login"]
+        .iter()
+        .map(|rc| home_dir.join(rc))
+        .collect();
+
+    home_dir.dfxvm_init().arg("--proceed").assert().success();
+
+    for rc in &rcs {
+        assert!(!rc.exists(), "{} should not exist", rc.display());
+    }
+}
+
+#[test]
+fn detects_zsh_by_shell_env_var() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+    server.expect_install_latest();
+
+    let home_zshenv = home_dir.join(".zshenv");
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("SHELL", "/bin/zsh")
+        .env("PATH", home_dir.join("nothing"))
+        .assert()
+        .success();
+
+    assert_eq!(
+        std::fs::read_to_string(home_zshenv).unwrap(),
+        posix_source()
+    );
+}
+
+#[test]
+fn detects_zsh_by_zsh_on_path() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    let home_zshenv = home_dir.join(".zshenv");
+
+    let bin = tempfile::Builder::new()
+        .prefix("dfxvm-integration-tests-bin")
+        .tempdir()
+        .unwrap();
+    // just has to exist
+    std::fs::write(bin.path().join("zsh"), "").unwrap();
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("PATH", bin.path())
+        .assert()
+        .success();
+
+    assert_eq!(
+        std::fs::read_to_string(home_zshenv).unwrap(),
+        posix_source()
+    );
+}
+
+#[test]
+fn does_not_detect_zsh_by_zdotdir_env_var() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    let zdotdir = home_dir.join("zdotdir");
+    std::fs::create_dir(&zdotdir).unwrap();
+
+    // let's even create a zshenv in there
+    let zdotdir_zshenv = zdotdir.join(".zshenv");
+    std::fs::write(&zdotdir_zshenv, FAKE_RC).unwrap();
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("ZDOTDIR", &zdotdir)
+        .env("PATH", home_dir.join("nothing"))
+        .assert()
+        .success();
+
+    assert_eq!(std::fs::read_to_string(&zdotdir_zshenv).unwrap(), FAKE_RC);
+}
+
+#[test]
+fn creates_zshenv() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    let home_zshenv = home_dir.join(".zshenv");
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("SHELL", "/bin/zsh")
+        .env("PATH", home_dir.join("nothing"))
+        .assert()
+        .success();
+
+    assert_eq!(
+        std::fs::read_to_string(home_zshenv).unwrap(),
+        posix_source()
+    );
+}
+
+#[test]
+fn updates_zshenv() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    let home_zshenv = home_dir.join(".zshenv");
+    std::fs::write(&home_zshenv, FAKE_RC).unwrap();
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("SHELL", "/bin/zsh")
+        .env("PATH", home_dir.join("nothing"))
+        .assert()
+        .success();
+
+    assert_eq!(
+        std::fs::read_to_string(&home_zshenv).unwrap(),
+        FAKE_RC.to_owned() + &posix_source()
+    );
+}
+
+#[test]
+fn creates_zdotdir_zshenv() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    let zdotdir = home_dir.join("zdotdir");
+    std::fs::create_dir(&zdotdir).unwrap();
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("SHELL", "/bin/zsh")
+        .env("ZDOTDIR", &zdotdir)
+        .env("PATH", home_dir.join("nothing"))
+        .assert()
+        .success();
+
+    let zdotdir_zshenv = zdotdir.join(".zshenv");
+    assert_eq!(
+        std::fs::read_to_string(zdotdir_zshenv).unwrap(),
+        posix_source()
+    );
+}
+
+#[test]
+fn updates_zdotdir_zshenv() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    let zdotdir = home_dir.join("zdotdir");
+    std::fs::create_dir(&zdotdir).unwrap();
+    std::fs::write(zdotdir.join(".zshenv"), FAKE_RC).unwrap();
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("SHELL", "/bin/zsh")
+        .env("ZDOTDIR", &zdotdir)
+        .env("PATH", home_dir.join("nothing"))
+        .assert()
+        .success();
+
+    let zdotdir_zshenv = zdotdir.join(".zshenv");
+    assert_eq!(
+        std::fs::read_to_string(zdotdir_zshenv).unwrap(),
+        FAKE_RC.to_owned() + &posix_source()
+    );
+}
+
+#[test]
+fn gets_zdotdir_by_calling_zsh() {
+    let home_dir = TempHomeDir::new();
+    // this test requires that zsh is callable.
+    if home_dir
+        .new_command("zsh")
+        .arg("-c")
+        .arg("true")
+        .output()
+        .is_err()
+    {
+        return;
+    }
+
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    let home_zshenv = home_dir.join(".zshenv");
+    let zdotdir = home_dir.join("my-zdot-dir");
+    let zdotdir_zshenv = zdotdir.join(".zshenv");
+
+    std::fs::create_dir(&zdotdir).unwrap();
+    std::fs::write(zdotdir.join(".zshenv"), FAKE_RC).unwrap();
+
+    let export_zdotdir = format!(
+        "export ZDOTDIR={}",
+        zdotdir.into_os_string().to_str().unwrap()
+    );
+    std::fs::write(home_zshenv, export_zdotdir).unwrap();
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("SHELL", "/bin/sh")
+        .assert()
+        .success();
+    //.stderr("abc");
+
+    assert_eq!(
+        std::fs::read_to_string(zdotdir_zshenv).unwrap(),
+        FAKE_RC.to_owned() + &posix_source()
+    );
+}
+
+#[test]
+fn ignores_empty_zdotdir_env_var() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("SHELL", "/bin/zsh")
+        .env("ZDOTDIR", "")
+        .env("PATH", home_dir.join("nothing"))
+        .assert()
+        .success();
+
+    let home_zshenv = home_dir.join(".zshenv");
+    assert_eq!(
+        std::fs::read_to_string(home_zshenv).unwrap(),
+        posix_source()
+    );
+}
+
+#[test]
+fn prefers_zdotdir_to_home_zshenv_if_neither_exist() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    // here ZDOTDIR is set, but neither $ZDOTDIR/.zshenv nor $HOME/.zshenv exist
+    // so we should create $ZDOTDIR/.zshenv
+    let zdotdir = home_dir.join("the-zdot-dir");
+    // dfxvm-init will not create $ZDOTDIR, so we have to
+    std::fs::create_dir(&zdotdir).unwrap();
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("SHELL", "/bin/zsh")
+        .env("ZDOTDIR", &zdotdir)
+        .assert()
+        .success();
+
+    let zdotdir_zshenv = zdotdir.join(".zshenv");
+    assert_eq!(
+        std::fs::read_to_string(zdotdir_zshenv).unwrap(),
+        posix_source()
+    );
+    assert!(!home_dir.join(".zshenv").exists());
+}
+
+#[test]
+fn prefers_zdotdir_to_home_zshenv_if_both_exist() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    // here ZDOTDIR is set, and both $ZDOTDIR/.zshenv and $HOME/.zshenv exist
+    // so we should update $ZDOTDIR/.zshenv
+    let zdotdir = home_dir.join("the-zdot-dir");
+    let zdotdir_zshenv = zdotdir.join(".zshenv");
+    let home_zshenv = home_dir.join(".zshenv");
+    std::fs::create_dir(&zdotdir).unwrap();
+    std::fs::write(zdotdir_zshenv, FAKE_RC).unwrap();
+    std::fs::write(&home_zshenv, "echo this is the home zshenv").unwrap();
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("SHELL", "/bin/zsh")
+        .env("ZDOTDIR", &zdotdir)
+        .assert()
+        .success();
+
+    let zdotdir_zshenv = zdotdir.join(".zshenv");
+    assert_eq!(
+        std::fs::read_to_string(zdotdir_zshenv).unwrap(),
+        FAKE_RC.to_owned() + &posix_source()
+    );
+    assert_eq!(
+        std::fs::read_to_string(home_zshenv).unwrap(),
+        "echo this is the home zshenv"
+    );
+}
+
+#[test]
+fn prefers_home_zshenv_if_zdotdir_zshenv_does_not_exist() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    // here ZDOTDIR is set, but $ZDOTDIR/.zshenv does not exist.
+    // $HOME/.zshenv does exist, so we should update that.
+
+    let zdotdir = home_dir.join("the-zdot-dir");
+    // dfxvm-init will not create $ZDOTDIR, so we have to
+    std::fs::create_dir(&zdotdir).unwrap();
+
+    let home_zshenv = home_dir.join(".zshenv");
+    std::fs::write(&home_zshenv, FAKE_RC).unwrap();
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .env("SHELL", "/bin/zsh")
+        .env("ZDOTDIR", &zdotdir)
+        .assert()
+        .success();
+
+    assert_eq!(
+        std::fs::read_to_string(home_zshenv).unwrap(),
+        FAKE_RC.to_owned() + &posix_source()
+    );
+    assert!(!zdotdir.join(".zshenv").exists());
+}
+
+#[test]
+fn confirmation_message_profile_scripts_modified() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    #[cfg(target_os = "macos")]
+    let expected_env_path = "$HOME/Library/Application Support/org.dfinity.dfx/env";
+    #[cfg(target_os = "linux")]
+    let expected_env_path = "$HOME/.local/share/dfx/env";
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .assert()
+        .success()
+        .stdout(contains("dfxvm is installed now."))
+        .stdout(contains(
+            "To get started you may need to restart your current shell",
+        ))
+        .stdout(contains(
+            "This would reload your PATH environment variable to include",
+        ))
+        .stdout(contains("the dfxvm bin directory"))
+        .stdout(contains("To configure your shell, run:"))
+        .stdout(contains(format!(r#"  source "{expected_env_path}""#)));
+
+    assert!(home_dir.installed_dfx_path("0.15.0").exists());
+    assert_eq!(home_dir.settings().read_default_version(), "0.15.0");
+}
+
+#[test]
+fn confirmation_message_profile_scripts_not_modified() {
+    let home_dir = TempHomeDir::new();
+    let server = ReleaseServer::new(&home_dir);
+
+    server.expect_install_latest();
+
+    #[cfg(target_os = "macos")]
+    let expected_env_path = "$HOME/Library/Application Support/org.dfinity.dfx/env";
+    #[cfg(target_os = "linux")]
+    let expected_env_path = "$HOME/.local/share/dfx/env";
+
+    home_dir
+        .dfxvm_init()
+        .arg("--proceed")
+        .arg("--no-modify-path")
+        .assert()
+        .success()
+        .stdout(contains("dfxvm is installed now."))
+        .stdout(contains(
+            "To get started you need the dfxvm bin directory in your PATH",
+        ))
+        .stdout(contains("This has not been done automatically"))
+        .stdout(contains("To configure your shell, run:"))
+        .stdout(contains(format!(r#"  source "{expected_env_path}""#)));
+
+    assert!(home_dir.installed_dfx_path("0.15.0").exists());
+    assert_eq!(home_dir.settings().read_default_version(), "0.15.0");
 }